% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDBF.R
\name{cleanDBF}
\alias{cleanDBF}
\title{Clean .DBF}
\usage{
cleanDBF(x, y = NULL)
}
\arguments{
\item{x}{input dataframe with data}

\item{y}{input dataframe with column names if missing, defaults to NULL}
}
\value{
\code{x} output dataframe
}
\description{
Function to clean .DBF files
}
