% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_plot.R
\name{contribution_plot}
\alias{contribution_plot}
\title{Construct contribution plot}
\usage{
contribution_plot(rawData, trainingData, obs, multistate = TRUE)
}
\arguments{
\item{rawData}{where obs is found}

\item{trainingData}{ADPCA training specifications}

\item{obs}{observation to plot}
}
\value{
difference.normalized
}
\description{
This function tests data for multistate AD-PCA
}
