% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingSpecs.R
\name{createTrainingSpecs}
\alias{createTrainingSpecs}
\title{Run Training Functions}
\usage{
createTrainingSpecs(data, testingDay, rollingWindowDays, alpha = 0.01,
  faultsToTriggerAlarm = 5)
}
\arguments{
\item{data}{Dataframe of process data}

\item{testingDay}{Date object, all dates before this date will be included in the training analysis}

\item{rollingWindowDays}{Number of observations used in the rolling window}

\item{alpha}{Accuracy of analysis, defaults to 0.01}

\item{faultsToTriggerAlarm}{Number of abnormal observations to be considered an alarm, defaults to five}
}
\description{
Using only the data from the start to the day before testing day, train the ADPCA model using functions from mvMonitoring
}
\examples{
trainingSpecHolder <- createTrainingSpecs(data = dataBR[[i]],
testingDay = testingDay,
trainObs = rollingWindowObs,
updateFreq = updateObs,
alpha = alphaN,
faultsToTriggerAlarm = faultsToTriggerAlarm)
}
