% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmerge.R
\name{fastmerge}
\alias{fastmerge}
\title{Merge two dataframes}
\usage{
fastmerge(d1, d2)
}
\arguments{
\item{d1}{Dataframe}

\item{d2}{Dataframe}
}
\description{
Merges two dataframes, both by column and row, and replaces unoccupied cells with 'NA'
}
