% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_alarmData.R
\name{graph_alarmData}
\alias{graph_alarmData}
\title{Graph alarm data}
\usage{
graph_alarmData(data, keyword = "", trueFaultTime = NULL)
}
\arguments{
\item{data}{compiled xts object}
}
\description{
Function to plot all varaibles in original .DBF file
}
