% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_rawData.R
\name{graph_rawData}
\alias{graph_rawData}
\title{Graph raw data}
\usage{
graph_rawData(data, keyword = "", trueFaultTime = NULL)
}
\arguments{
\item{data}{compiled xts object}

\item{keyword}{string for filename}
}
\description{
Function to plot all varaibles in original .DBF file
}
