% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadandcleanDBF.R
\name{loadandcleanDBF}
\alias{loadandcleanDBF}
\title{Import and Prepare Data}
\usage{
loadandcleanDBF(dataLocation, testingDay, nDays)
}
\arguments{
\item{dataLocation}{String excluding ending backslashes of LogAllData folder}

\item{testingDay}{Date object with the last date to be included}

\item{nDays}{Number of days of data to compile}
}
\value{
rawData, Dataframe including all compiled data
}
\description{
This function compiles raw data and column names and elimates extraneous columns
}
\examples{
dataLocation <- "MP SB-MBR Data\\\\"
testingDay <- as.Date("2017 08 31", format = "\%Y \%m \%d")
nDays <- 30
rawData <- loadandcleanDBF(dataLocation, testingDay, nDays)
}
