% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDBF.R
\name{loadDBF}
\alias{loadDBF}
\title{Load all original data files}
\usage{
loadDBF(dataLocation, endDay, nDays, folder = "LogAllData")
}
\arguments{
\item{dataLocation}{String excluding ending backslashes of LogAllData folder}

\item{endDay}{Date object with the last date to be included}

\item{nDays}{Number of days of data to compile}

\item{folder}{LogAllData or Log60sec}
}
\value{
compiledData, Dataframe with all original data from .DBF files.
}
\description{
This function loops through all filenames within the date range, checks that the file exists, and if it does exist, is imported and compiled to a dataframe
}
\examples{
rawData <- loadDBF(dataLocation, testingDay, nDays)
}
