% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistate_test.R
\name{multistate_test}
\alias{multistate_test}
\title{Test based on multistate data}
\usage{
multistate_test(data, trainingSpec_ls, testingDay, faultsToTriggerAlarm)
}
\value{
alarms.xts
}
\description{
This function tests data for multistate AD-PCA
}
