% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistate_train.R
\name{multistate_train}
\alias{multistate_train}
\title{Train based on multistate data}
\usage{
multistate_train(rawData, vars, stateVars, testingDay, rollingWindowDays,
  alphaN, faultsToTriggerAlarm)
}
\value{
alarms.xts
}
\description{
This function trains a multistate AD-PCA based on data
}
