% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLoad.R
\name{packageLoad}
\alias{packageLoad}
\title{Install and load libraries}
\usage{
packageLoad(packName)
}
\arguments{
\item{packName}{package name as a character string, e.g. "quantmod"}
}
\description{
This function checks to see if a package is already installed before installing and loading it.
}
\examples{
packageLoad(ADPCA)
}
