% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDates.R
\name{setDates}
\alias{setDates}
\title{Parse dates}
\usage{
setDates(startDay, endDay)
}
\arguments{
\item{startDay}{First day of data}

\item{endDay}{Testing day of data}
}
\value{
List of integers
}
\description{
Take the start and end date and turn them into integers that can be iterated through
}
