% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateGenerator.R
\name{stateGenerator}
\alias{stateGenerator}
\title{State Generator}
\usage{
stateGenerator(data, stateVars, minObs = round((2 *
  length(which(!(colnames(data) \%in\% stateVars))))^2/2), testingDay,
  rollingWindowDays = 0)
}
\arguments{
\item{data}{Dataframe with column names}

\item{stateVars}{Vector of character strings to identify state variables by their column names in \code{data}}

\item{minObs}{Number of minimum observations required to define a "state." Defaults to statistically required p^2/2 where p is the number of process variables (including lag)}

\item{testingDay}{Date object of the day to be evaluated for faults.}

\item{rollingWindowDays}{Integer with the number of days of data to be included in the training window.}
}
\value{
A list of xts objects with data, a merged "labelCol," and separated by operating state
}
\description{
Add a column to the dataset with a unique indentifier for the operating conditions
}
\examples{
dataBR <- stateGenerator(data = dataBR, stateVars = stateVarsBR, minObs = rollingWindowObs)

}
