% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNewObs.R
\name{testNewObs}
\alias{testNewObs}
\title{Test observations for abnormal conditions}
\usage{
testNewObs(data, trainingSpecs, testingDay, faultsToTriggerAlarm = 5)
}
\arguments{
\item{data}{xts object}

\item{trainingSpecs}{object returned from createTrainingSpecs()}

\item{testingDay}{Date object of testing observations}

\item{faultsToTriggerAlarm}{Defaults to 5}
}
\description{
This function compares distance of new observations to the subspace identified by ADPCA in trainingSpecs() by calling mspMonitor() and mspWarning() from mvMonitoring package
}
