% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniquenessCheck.R
\name{uniquenessCheck}
\alias{uniquenessCheck}
\title{Uniqueness Check}
\usage{
uniquenessCheck(data, N = 20)
}
\arguments{
\item{data}{An xts object}

\item{N}{Number of unique observations required. Defaults to 20.}
}
\description{
Check for changing process variables
}
