% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistate_train.R
\name{multistate_train}
\alias{multistate_train}
\title{Train based on multistate data}
\usage{
multistate_train(rawData, vars, stateVars, testingDay, rollingWindowDays,
  alphaN, faultsToTriggerAlarm, statistic = "T2", by = "Time")
}
\value{
alarms.xts
}
\description{
This function trains a multistate AD-PCA based on data. Data are separated into a subsystem based on the variables specified. A label column is created by concatenating all of the state variables (which are integers) into a unique (string) identifier. If there are sufficient instances of a particular identifier (by default p^2/2 where p is the number of process variables including lag), then the state is included. createTrainingSpecs() function will calculate the principal components used for testing.
}
